# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Studio Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users (admin only)
      parameters:
        - name: plan
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/grant-vip:
    post:
      operationId: grantVip
      tags: [admin]
      summary: Grant VIP access to user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: VIP granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        plan:
          type: string
          enum: [free, pro, enterprise, vip, admin]
        isAdmin:
          type: boolean
        isVip:
          type: boolean
        projectCount:
          type: integer
        buildsThisMonth:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - plan
        - isAdmin
        - isVip
        - projectCount
        - buildsThisMonth
        - createdAt
    UpdateUserRequest:
      type: object
      properties:
        plan:
          type: string
          enum: [free, pro, enterprise, vip]
        isAdmin:
          type: boolean
...