# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Studio Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserPlan } from "./userPlan";

export interface User {
  id: string;
  username: string;
  email?: string;
  plan: UserPlan;
  isAdmin: boolean;
  isVip: boolean;
  projectCount: number;
  buildsThisMonth: number;
  createdAt: Date;
}